Upgrade to Quarkus 3.7.1 (#26736)
Closes #26701
Closes #23854
