+++ pom.xml
@@ -45,8 +45,8 @@
         <jboss.snapshots.repo.id>jboss-snapshots-repository</jboss.snapshots.repo.id>
         <jboss.snapshots.repo.url>https://s01.oss.sonatype.org/content/repositories/snapshots/</jboss.snapshots.repo.url>
 
-        <quarkus.version>3.7.0.CR1</quarkus.version>
-        <quarkus.build.version>3.7.0.CR1</quarkus.build.version>
+        <quarkus.version>3.7.1</quarkus.version>
+        <quarkus.build.version>3.7.1</quarkus.build.version>
 
         <project.build-time>${timestamp}</project.build-time>
