+++ quarkus/runtime/src/main/resources/application.properties
@@ -7,7 +7,6 @@ quarkus.banner.enabled=false
 
 # Disable health checks from extensions, since we provide our own (default is true)
 quarkus.health.extensions.enabled=false
-quarkus.datasource.health.enabled=false
 
 # Disable http metrics binder as URL parameters are only shown with placeholders for '/resource' URLs, but not
 # for '/admin' and '/realms'. Neither the IDs of entities nor the realm name should be part of the metric names
