+++ testsuite/utils/pom.xml
@@ -266,6 +266,7 @@
         <dependency>
             <groupId>mysql</groupId>
             <artifactId>mysql-connector-java</artifactId>
+            <version>${mysql-jdbc.version}</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
